Title: 사각형 채우기 3, Time: 6ms, Memory: 0MB, Status: Wrong Answer - CodeTree

+++ 240822/사각형 채우기 3/rectangle-fill-3.cpp
@@ -1,7 +1,7 @@
 #include <iostream>
 using namespace std;
 
-int dp[1001];
+long long dp[1001];
 
 int main() {
     // 여기에 코드를 작성해주세요.
